test: modify and add tests for scanner
Modified all the existing tests for scanner related code so that it fits the new UI implemented. Fixed some old tests that weren't very stable and added new tests to test the new functionalities introduced with the new UI (checking formatting for date text boxes, checking that the new screen does proper error handling, checking that when scanned it shows the small pop up and then expands to the bigger pop up when clicked, etc...)

Some utils functions have still not been tested as they require some refactoring and compatibility with other classes that are still to be modified.