feat: modify UI/validation logic for AddFoodItemScreen to match FoodInputContent
Modified the UI and validation/error handling logic in AddFoodItemScreen so that it matches FoodInputContent and makes it easier for users to understand their errors and change them. Also added padding and spacing so that the screen looks nice in all phones it runs in (no matter the size). Also refactored some utility code used in FoodInputContent so that it is now in the utils class DateUtils.kt (used to be Time.kt) to avoid code duplication and make the structure of the project more coherent.