test: add tests for AddFoodItemScreen class and format project
Added/Modified existing tests for AddFoodItemScreen so that they match the new UI and functionalities. This led to modifying the DropdownFields composable in order to make testing easier, and also led to further modification in AddFoodItemScreen after some bugs and odd functionalities were caught in tests (such as the error messages only showing up after having interacted with that textbox at least once). Modifications to the EndToEndM1Test were also necessary as the functionality had changed and one of the tests no longer passed (since there is no longer an error pop up). Finally, also formatted the code with ktfmtFormat to ensure ktfmt check passes on the CI.