Refactor Utility Functions, Update AddFoodItemScreen, and Improve/Make New Tests for DateUtils and AddFoodItemScreen (#148)
* feat: modify UI/validation logic for AddFoodItemScreen to match FoodInputContent

Modified the UI and validation/error handling logic in AddFoodItemScreen so that it matches FoodInputContent and makes it easier for users to understand their errors and change them. Also added padding and spacing so that the screen looks nice in all phones it runs in (no matter the size). Also refactored some utility code used in FoodInputContent so that it is now in the utils class DateUtils.kt (used to be Time.kt) to avoid code duplication and make the structure of the project more coherent.

* test: add tests for DateUtils.kt class

Added tests for all the new functions in DateUtils.kt and modified existing ones so they match the new way of formatting dates and error handling (with fixing the format and adding the slashes on its own). Also ended up fixing issues in DateUtils.kt that were caught in the new tests (and fixed small issue in Scanner due to a careless refactor in the previous commit)

* test: add tests for AddFoodItemScreen class and format project

Added/Modified existing tests for AddFoodItemScreen so that they match the new UI and functionalities. This led to modifying the DropdownFields composable in order to make testing easier, and also led to further modification in AddFoodItemScreen after some bugs and odd functionalities were caught in tests (such as the error messages only showing up after having interacted with that textbox at least once). Modifications to the EndToEndM1Test were also necessary as the functionality had changed and one of the tests no longer passed (since there is no longer an error pop up). Finally, also formatted the code with ktfmtFormat to ensure ktfmt check passes on the CI.

* feat: make AddFoodItemScreen scrollable to pass CI tests and improve functionality in phones with smaller screens

Made AddFoodItemScreen scrollable so that in case it doesnt fully fit into the screen due to the screen being too small (like in the CI), you can scroll around and find all the components. Also modified the tests so that they scroll to the component before checking to ensure they are indeed displayed/

* fix: improve scrolling by adding tags to each item so tests pass

Added 'tags' to each of the items displayed in AddFoodItemScreen so that in the tests we can scroll to those tags and ensure the CI can correctly find the nodes and scroll to them (and then check that they are displayed)

* test: fix issue in testSubmitButtonWithInvalidForm where it doesn't scroll to component before checking that it is displayed

* chore: refactor and extract duplicated code to decrease code duplication in AddFoodItemScreen

* fix: fix errors related to displaying error messages introduced in the last refactor

* docs: add documentation to classes and methods in AddFoodItem.kt and DateUtils.kt

---------

Co-authored-by: Mateus Vital Nabholz <mateus.vitaln@icloud.com>