Feat/152 general bug fixes and UI improvements on overview and household creation screens (#158)
* fix: fix the progress bar and creation screen buttons

Progress bar is now full when green and small when red. Creation screen save and cancel buttons have been adapted to fit inside small screens.

* fix: fix items not refreshing when adding a new item in the overview

The issue was that the selectedHousehold was not updated correctly in the viewModel, this fixes it but our HouseholdViewModel has major issues and needs an overhaul. Tests needed to be adapted.

* fix: moved hard coded colors to Color.kt