fix: fix issue with main in which foodItem was selected from HouseholdViewModel instead of ListFoodItemsViewModel in overview.
removed function and variable related to selectedFoodItem in HouseholdViewModel and instead had it be called from ListFoodItemsViewModel in Overview.kt for IndividualFoodItem.kt. This led to also having to modify IndividualFoodItem.kt and its tests to make it compatible. The rest (and one that has affected most lines) of the modifications were the result of formatting the code using ktfmtFormat on the project