Add edit food item screen (#162)
* feat: implement EditFoodItemScreen and extract useful validation functions into a separate class so they can be reused by all screens that do such validation

Fully implemented EditFoodItemScreen which will be later integrated into the project once IndividualFoodItemScreen in on Main and also extracted all the different validation functions used in AddFoodItem.kt into a separate class so it can be used by all composables that perform such validation

* feat: implement way to change amount of food (not the unit) in EditFoodItem.kt, implement function that modifies foodlist in household so if a foodItem is edited, the household gets properly updated, and made tests for EditFoodItemScreen and the new HouseholdViewModel function

* fix: fix issue with main in which foodItem was selected from HouseholdViewModel instead of ListFoodItemsViewModel in overview.

removed function and variable related to selectedFoodItem in HouseholdViewModel and instead had it be called from ListFoodItemsViewModel in Overview.kt for IndividualFoodItem.kt. This led to also having to modify IndividualFoodItem.kt and its tests to make it compatible. The rest (and one that has affected most lines) of the modifications were the result of formatting the code using ktfmtFormat on the project

---------

Co-authored-by: Mateus Vital Nabholz <mateus.vitaln@icloud.com>