feat: make barcode scanner related classes fit with the new app architecture
This commit contains the following: modifying BarcodeScannerViewModel so that it also does foodfacts operations. Update Scanner.kt so that it uses the new BarcodeScannerViewModel. Modify FoodInputContent.kt so that it uses FoodItemViewModel. Update FoodItemViewModel so that it does extra checks and handles operations properly when done from the scanner screen (since its slightly different from the Add and Edit foodItem cases)