278 refactor barcode scanners and camera elements to work with new architecture (#282)
* feat: make barcode scanner related classes fit with the new app architecture

This commit contains the following: modifying BarcodeScannerViewModel so that it also does foodfacts operations. Update Scanner.kt so that it uses the new BarcodeScannerViewModel. Modify FoodInputContent.kt so that it uses FoodItemViewModel. Update FoodItemViewModel so that it does extra checks and handles operations properly when done from the scanner screen (since its slightly different from the Add and Edit foodItem cases)

* fix: fix MainActivity.kt so that it uses the new CameraPermissionHandler and new BarcodeScannerScreen (fix what they were given as parameters in MainActivity)

* feat: update OpenFoodFactsRepository to contain the old viewModel functions.

This repository now replaces the ViewModel

* feat: add hilt FoodFactRepo injection module

---------

Co-authored-by: Mateus Vital Nabholz <mateus.vitaln@icloud.com>
Co-authored-by: alextheprophecy <alexandrebourgoin23@gmail.com>
Co-authored-by: Robin Herberich <99408085+ElMaestroZ@users.noreply.github.com>