fix: manage household deletion as discussed in group and make firebase requests asynchronous
Made it so only the last member to leave a household truly deletes it. Else only their list of uids gets modified and the other can still use it as normal. Also made all firebase operation asynchronous (using onSuccess and onFailure listeners) so that the app works offline with proper error handling and rollbacks